# go/APIvidgm/docs/swagger.yaml
definitions:
  models.CDKey:
    properties:
      id:
        type: integer
      key:
        type: string
      order_item:
        $ref: '#/definitions/models.OrderItem'
      order_item_id:
        type: integer
      state:
        description: available, reserved, sold
        type: string
      user_id:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
  models.Cart:
    properties:
      id:
        type: integer
      total_price:
        type: number
      user_id:
        type: integer
      videogames:
        items:
          $ref: '#/definitions/models.CartItem'
        type: array
    type: object
  models.CartItem:
    properties:
      cart_id:
        type: integer
      id:
        type: integer
      is_physical:
        type: boolean
      quantity:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
  models.DigOrder:
    properties:
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        description: pending, completed, canceled
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  models.Direction:
    properties:
      city:
        type: string
      id:
        type: integer
      number:
        type: string
      state:
        type: string
      street:
        type: string
      user_id:
        type: integer
      zip:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Label:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.OrderItem:
    properties:
      cdkey:
        items:
          $ref: '#/definitions/models.CDKey'
        type: array
      id:
        type: integer
      orderOwnerID:
        type: integer
      orderOwnerType:
        description: '"phy_order" o "dig_order"'
        type: string
      quantity:
        type: integer
      videogame:
        $ref: '#/definitions/models.VideoGame'
      videogame_id:
        type: integer
    type: object
  models.PhyOrder:
    properties:
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/models.OrderItem'
        type: array
      status:
        description: pending, completed, canceled
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  models.Platform:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      id:
        type: integer
      modstatus:
        description: pending, approved, rejected
        type: string
      rating:
        type: integer
      user_id:
        type: integer
      username:
        type: string
      videogame_id:
        type: integer
      videogame_name:
        type: string
    type: object
  models.User:
    properties:
      cart:
        $ref: '#/definitions/models.Cart'
      dig_orders:
        items:
          $ref: '#/definitions/models.DigOrder'
        type: array
      directions:
        items:
          $ref: '#/definitions/models.Direction'
        type: array
      id:
        type: integer
      mail:
        type: string
      password:
        type: string
      phy_orders:
        items:
          $ref: '#/definitions/models.PhyOrder'
        type: array
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      role:
        type: string
      username:
        type: string
      wishlist:
        $ref: '#/definitions/models.WishList'
    type: object
  models.VideoGame:
    properties:
      developer:
        type: string
      dig_stock:
        type: integer
      genre:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      image_url:
        type: string
      label:
        items:
          $ref: '#/definitions/models.Label'
        type: array
      min_req:
        type: string
      pegi:
        type: integer
      phy_stock:
        type: integer
      platform:
        items:
          $ref: '#/definitions/models.Platform'
        type: array
      price:
        type: number
      rating:
        type: number
      release_year:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/models.Review'
        type: array
      sum_rating:
        type: integer
      synopsis:
        type: string
      title:
        type: string
    type: object
  models.WishList:
    properties:
      id:
        type: integer
      user_id:
        type: integer
      videogames:
        items:
          $ref: '#/definitions/models.VideoGame'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Shit documentation for shit API.
  termsOfService: http://swagger.io/terms/
  title: VideoGames API
  version: "1.0"
paths:
...
  /directions/{dirID}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific delivery direction by its ID. Requires authentication
        and that the direction belongs to the authenticated user.
      parameters:
      - description: Direction ID
        in: path
        name: dirID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Direction deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Direction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user's delivery direction
      tags:
      - Directions
    get:
      consumes:
      - application/json
      description: Retrieves a specific delivery direction by its ID. Requires authentication
        and that the direction belongs to the authenticated user.
      parameters:
      - description: Direction ID
        in: path
        name: dirID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Direction'
        "400":
          description: Direction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a delivery direction by ID
      tags:
      - Directions
...
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves a user by ID along with their cart information. Requires
        authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - Users
...
  /users/{id}/directions:
    get:
      consumes:
      - application/json
      description: Retrieves all saved delivery directions for a specific user. Requires
        authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user's delivery directions
          schema:
            items:
              $ref: '#/definitions/models.Direction'
            type: array
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user's delivery directions
      tags:
      - Directions
    post:
      consumes:
      - application/json
      description: Creates and associates a new delivery direction with the specified
        user. Requires authentication and the same user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: New delivery direction data
        in: body
        name: direction
        required: true
        schema:
          $ref: '#/definitions/models.Direction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Direction'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add a new delivery direction for a user
      tags:
      - Directions
  /users/{id}/directions/{dirID}:
    put:
      consumes:
      - application/json
      description: Updates the details of an existing delivery direction. Requires
        authentication and that the direction belongs to the authenticated user.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Direction ID
        in: path
        name: dirID
        required: true
        type: integer
      - description: Updated delivery direction data
        in: body
        name: direction
        required: true
        schema:
          $ref: '#/definitions/models.Direction'
      produces:
      - application/json
      responses:
        "200":
          description: Direction updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: User or direction not found
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Unauthorized or insufficient permissions
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user's delivery direction
      tags:
      - Directions
...
swagger: "2.0"